Added a simple event handler for Sheet.Change (endless loop!).
When trying to write the address of a changed cell to B2, we see that the
handler gets called again when cell B2 changes. This leads to an endless loop!
When using Sheet.Change, we must proceed with caution as it can lead to bad
performance (repeated calculations, ...) or, like in this case, an endless loop.